stabilize D4
Unfortunately the D4 response was not "stable" due to an uninitialized buffer.
Fixed that issue with fill(0). As an alternative man could switch node js 6 or higer and use Buffer.alloc(x) instead of new Buffer(x).
Added handling of write request.
Addes some more logging. (But it is disabled / commented out)